+++ server/internal/dto/book_category.go
@@ -3,13 +3,13 @@ package dto
 import dbmodels "github.com/scul0405/my-shop/server/db/models"
 
 type BookCategoryDTO struct {
-	ID   int64  `json:"id"`
+	ID   uint64 `json:"id"`
 	Name string `json:"name"`
 }
 
 func (b *BookCategoryDTO) ToModel() *dbmodels.BookCategory {
 	return &dbmodels.BookCategory{
-		ID:   b.ID,
+		ID:   int64(b.ID),
 		Name: b.Name,
 	}
 }
